Add support to render the active item icon on the Host
